% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_agreement_check.R
\name{gtfs_agreement_check}
\alias{gtfs_agreement_check}
\title{GTFS Agreement Check}
\usage{
gtfs_agreement_check(gtfs)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}
}
\value{
Either a table of trip_ids found in trips but not stop_times (or vice versa) or a message
}
\description{
Checks for agreement between stop_times.txt and trips.txt
}
