% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_calendar_info.R
\name{gtfs_calendar_info}
\alias{gtfs_calendar_info}
\title{GTFS Calendar Information}
\usage{
gtfs_calendar_info(gtfs, write_to_gtfs = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{write_to_gtfs}{Logical, whether to return the calendar information
as a table or add it to the GTFS as \code{gtfs$calendar_info}}
}
\value{
If \code{write_to_gtfs} is TRUE, a GTFS stored as a list. If FALSE,
a table
}
\description{
GTFS Calendar Information
}
