% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_merge.R
\name{gtfs_merge}
\alias{gtfs_merge}
\title{Merge GTFS files}
\usage{
gtfs_merge(gtfs1, gtfs2)
}
\arguments{
\item{gtfs1}{A GTFS file stored as a list}

\item{gtfs2}{A GTFS file stored as a list}
}
\value{
A GTFS file reflecting services stored in \code{gtfs1} and \code{gtfs2}
}
\description{
\code{gtfs_merge()} combines services from two distinct GTFS files. To read GTFS
files into R, try using \code{\link{read_mbta_gtfs}} or
\code{\link[tidytransit]{read_gtfs}}.
}
