% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_removal_check.R
\name{gtfs_removal_check}
\alias{gtfs_removal_check}
\title{Check for unused GTFS components}
\usage{
gtfs_removal_check(gtfs)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}
}
\value{
Either a list containing unused stops, shapes, routes, or \code{service_id}s or a message.
}
\description{
Check for unused GTFS components
}
\seealso{
Other gtfs cleaning functions: 
\code{\link{gtfs_remove_unused}}
}
\concept{gtfs cleaning functions}
