% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_remove_functions.R
\name{gtfs_remove_unused}
\alias{gtfs_remove_unused}
\alias{remove_unused_stops}
\alias{remove_unused_shapes}
\alias{remove_unused_routes}
\alias{remove_unused_service}
\title{GTFS cleaning functions}
\usage{
remove_unused_stops(gtfs, retain_stops = FALSE)

remove_unused_shapes(gtfs, retain_shapes = FALSE)

remove_unused_routes(gtfs, retain_routes = FALSE)

remove_unused_service(gtfs, retain_service = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{retain_stops}{Whether to retain the unused stops (in \code{gtfs$stops_unused})}

\item{retain_shapes}{Whether to retain the unused shapes (in \code{gtfs$shapes_unused})}

\item{retain_routes}{Whether to retain the unused routes (in \code{gtfs$routes_unused})}

\item{retain_service}{Whether to retain the unused \code{service_id}s (in \code{gtfs$calendar_unused})}
}
\value{
A GTFS file stored as a list, without unused stops.
}
\description{
GTFS cleaning functions
}
\concept{gtfs cleaning functions}
