% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabgs_19.R
\docType{data}
\name{mabgs_19}
\alias{mabgs_19}
\title{2019 ACS MA Block Group Demographics}
\format{
A data frame with 4985 rows and 14 variables:
\describe{
\item{GEOID}{Character, Block Group GEOID}
\item{pop}{Population}
\item{minopop}{Population meeting Title VI Racial or Ethnic minority definition}
\item{tothh}{Number of households}
\item{li45hh}{Number of households with household income under $45k/year}
\item{li50hh}{Number of households with household income under $50k/year}
\item{limenghh}{Number of households with limited English proficiency}
\item{MHHI}{Calculated median household income}
\item{lownocarhh}{Number of households with less than 0.5 cars per person}
\item{foreignbornres}{Number of foreign-born residents}
\item{minopct}{Percent of population meeting Titile VI Racial or Ethnic minority definition}
\item{li45pct}{Percent of households with household income under $45k/year}
\item{li50pct}{Percent of households with household income under $50k/year}
\item{lownocarpct}{Percent of households with less than 0.5 cars per person}
}
}
\usage{
mabgs_19
}
\description{
Contains frequently used demographic information for MA block groups from the 2019 ACS 5-year estimates.
}
\keyword{datasets}
