% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbta_service_area.R
\docType{data}
\name{mbta_service_area}
\alias{mbta_service_area}
\title{MBTA Service Area}
\format{
A data frame with 351 rows and 3 variables
\describe{
\item{muni_name}{Name of MA municipality}
\item{mbta65}{Logical, whether the municipality is served by MBTA subway/bus}
\item{mbta175}{Logical, whether the municipality is served by MBTA}
}
}
\usage{
mbta_service_area
}
\description{
Table contains one row per MA municipality.
}
\keyword{datasets}
