% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{opmi_cols}
\alias{opmi_cols}
\title{OPMI Colors}
\usage{
opmi_cols(...)
}
\arguments{
\item{...}{One, or multiple colors from the OPMI color palette
("orange", "petrolblue", "coralred", "seafoam", "skyblue", "deeppurple",
"cream", or "midgrey")
or MBTA color scheme
("rl", "ol", "gl", "bl", "bus", "ferry", "cr", "sl", "ride").
If no colors are provided, the entire list of colors will be returned.}
}
\value{

}
\description{
A function to provide hex values for colors in the OPMI color palette
or colors of MBTA services.
}
\examples{
opmi_cols()
opmi_cols("coralred")
opmi_cols("rl", "ol")
}
