% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{opmi_pal}
\alias{opmi_pal}
\title{OPMI Palette Chooser}
\usage{
opmi_pal(
  palette = "main",
  reverse = FALSE,
  discrete = TRUE,
  mono = FALSE,
  monocol = "orange",
  ...
)
}
\arguments{
\item{palette}{One of "main", "diverging", "diverging2", or "rt"}

\item{reverse}{Logical, whether to reverse palette}

\item{discrete}{Logical, whether palette is for discrete scale}

\item{mono}{Logical, whether to use a single-color palette}

\item{monocol}{If \code{mono} = TRUE, which color to use for palette}

\item{...}{Additional arguments passed to \code{colorRampPalette}}
}
\value{
A function that can be used by \code{scale_fill_opmi} and \code{scale_color_opmi}
}
\description{
Creates a function that can be used by \code{scale_fill_opmi} and \code{scale_color_opmi}.
Can also be used to create manual color/fill palettes for \code{scale_color_manual}
and \code{scale_fill_manual}.
}
