% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_month.R
\name{plot_month}
\alias{plot_month}
\title{Create Plot Month}
\usage{
plot_month(month_field)
}
\arguments{
\item{month_field}{Date field in data frame}
}
\value{
Factor field in data frame
}
\description{
Create Plot Month
}
