% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_remove_functions.R
\name{remove_unused_routes}
\alias{remove_unused_routes}
\title{GTFS cleaning functions}
\usage{
remove_unused_routes(gtfs, retain_routes = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{retain_routes}{Whether to retain the unused routes (in \code{gtfs$routes_unused})}
}
\value{
A GTFS file stored as a list, without unused routes.
}
\description{
GTFS cleaning functions
}
\seealso{
Other gtfs cleaning functions: 
\code{\link{remove_unused_service}()},
\code{\link{remove_unused_shapes}()},
\code{\link{remove_unused_stops}()}
}
\concept{gtfs cleaning functions}
