% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_remove_functions.R
\name{remove_unused_service}
\alias{remove_unused_service}
\title{GTFS cleaning functions}
\usage{
remove_unused_service(gtfs, retain_service = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{retain_service}{Whether to retain the unused \code{service_id}s (in \code{gtfs$calendar_unused})}
}
\value{
A GTFS file stored as a list, without unused \code{service_id}s
}
\description{
GTFS cleaning functions
}
\seealso{
Other gtfs cleaning functions: 
\code{\link{remove_unused_routes}()},
\code{\link{remove_unused_shapes}()},
\code{\link{remove_unused_stops}()}
}
\concept{gtfs cleaning functions}
