% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_remove_functions.R
\name{remove_unused_shapes}
\alias{remove_unused_shapes}
\title{GTFS cleaning functions}
\usage{
remove_unused_shapes(gtfs, retain_shapes = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{retain_shapes}{Whether to retain the unused shapes (in \code{gtfs$shapes_unused})}
}
\value{
A GTFS file stored as a list, without unused shapes.
}
\description{
GTFS cleaning functions
}
\seealso{
Other gtfs cleaning functions: 
\code{\link{remove_unused_routes}()},
\code{\link{remove_unused_service}()},
\code{\link{remove_unused_stops}()}
}
\concept{gtfs cleaning functions}
