% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_remove_functions.R
\name{remove_unused_stops}
\alias{remove_unused_stops}
\title{GTFS cleaning functions}
\usage{
remove_unused_stops(gtfs, retain_stops = FALSE)
}
\arguments{
\item{gtfs}{A GTFS file stored as a list}

\item{retain_stops}{Whether to retain the unused stops (in \code{gtfs$stops_unused})}
}
\value{
A GTFS file stored as a list, without unused stops.
}
\description{
GTFS cleaning functions
}
\seealso{
Other gtfs cleaning functions: 
\code{\link{remove_unused_routes}()},
\code{\link{remove_unused_service}()},
\code{\link{remove_unused_shapes}()}
}
\concept{gtfs cleaning functions}
