% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_ridership.R
\docType{data}
\name{rt_ridership}
\alias{rt_ridership}
\title{MBTA Rapid Transit Ridership}
\format{
A data frame with 36 rows and 4 columns.
\describe{
\item{rating}{character, Applicable fall rating}
\item{route_id}{Factor, Route ID of service}
\item{day_type_name}{Character, day type of service}
\item{tot_ons}{Numeric, average total boardings of service}
}
}
\source{
\url{https://mbta-massdot.opendata.arcgis.com/datasets/MassDOT::mbta-rail-ridership-by-time-period-season-route-line-and-stop/explore}
}
\usage{
rt_ridership
}
\description{
Dataset showing average fall ridership on Red, Orange, Green, and Blue Lines.
}
\keyword{datasets}
