% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_opmi}
\alias{theme_opmi}
\title{OPMI Theme}
\usage{
theme_opmi(base_size = 12, base_font = "Open Sans", base_family = "opensans")
}
\arguments{
\item{base_size}{Numeric font size}

\item{base_font}{Name of font to use (only need to specify if you would like
a font that is not included in base R and not included in the OPMI theme)}

\item{base_family}{Name of font family to use (only need to specify if you
would like a font that is not included in base R and not included in the
OPMI theme). For a base R font, use "sans", "serif", or "mono".}
}
\value{
A theme that can be used in ggplots.
}
\description{
Provides a ggplot2 theme that is consistent with the OPMI MS Office theme.
}
